refactor: util 패키지 생성, const 키 변경
