+++ app/src/main/java/com/kiwi/kiwitalk/di/ApplicationModule.kt
@@ -8,7 +8,7 @@ import com.google.android.gms.auth.api.signin.GoogleSignIn
 import com.google.android.gms.auth.api.signin.GoogleSignInClient
 import com.google.android.gms.auth.api.signin.GoogleSignInOptions
 import com.kiwi.kiwitalk.AppPreference
-import com.kiwi.kiwitalk.Const
+import com.kiwi.kiwitalk.util.Const
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
