+++ app/src/main/java/com/kiwi/kiwitalk/util/Const.kt
@@ -1,13 +1,12 @@
-package com.kiwi.kiwitalk
+package com.kiwi.kiwitalk.util
 
 object Const {
     const val LOGIN_ID_KEY = "login_history_key"
     const val LOGIN_NAME_KEY = "login_name"
     const val LOGIN_URL_KEY = "login_url"
     const val EMPTY_STRING = ""
-    const val SERVER_CLIENT_ID_KEY = "SERVER_CLIENT_ID"
-    const val MAP_KEY_KEYWORD = "keyword"
-    const val MAP_KEY_COUNTRY = "country"
+    const val CHAT_KEYWORDS = "keywords"
+    const val CHAT_ADDRESS = "address"
     const val BASE_URL = "https://dapi.kakao.com/"
     const val KAKAO_AUTH = "Authorization"
     const val ADDRESS_ERROR = "현재 좌표에서 주소를 가져올 수 없습니다."
